+++ src/face/database/FaceDatabase.h
@@ -22,7 +22,7 @@ public:
 	bool Save(const char* path) const;
 	int Delete(const std::string& name);
 	int64_t Insert(const std::vector<float>& feat, const std::string& name);
-	int64_t QueryTop(const std::vector<float>& feat, QueryResult& query_result) const;
+	int QueryTop(const std::vector<float>& feat, QueryResult& query_result) const;
 
 
 private:
